copying replication package into repository
I made this replication package for internal use on a project with Dr. Jesse Bruhn and Dr. Thea How Choon on the effects of national-level political events on the speech used in city council meetings. Its purpose was to make my code flexible and replicable, since I was leaving the project at the end of the semester.